move content myflight to home
