+++ App/src/hubs/MainStack.js
@@ -4,7 +4,6 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import Home from '../screen/Home/Home';
 import Login from '../screen/Login/Login';
 import SignIn from '../screen/SignIn/SignIn';
-import MyFlight from '../screen/MyFlights/MyFlight';
 export default function mainStack() {
   const Stack = createNativeStackNavigator();
 
@@ -18,7 +17,6 @@ export default function mainStack() {
         <Stack.Screen name="Login" component={Login} />
         <Stack.Screen name="Home" component={Home} />
         <Stack.Screen name="SignIn" component={SignIn} />
-        <Stack.Screen name="MyFlight" component={MyFlight} />
       </Stack.Navigator>
     </NavigationContainer>
   );
